renaming azure_rm_virtualmachine_extension
