+++ test/integration/targets/azure_rm_virtualmachineextension/aliases
