+++ test/integration/targets/azure_rm_virtualmachineextension/meta/main.yml
